requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
Flask==2.0.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.0
matplotlib==3.4.2
mplfinance==0.12.7a17
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
polygon-api-client==0.1.11
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
urllib3==1.26.4
websocket-client==0.59.0
websockets==9.0.1
Werkzeug==2.0.1
yfinance==0.1.59
